Add files via upload
A step-by-step explanation of how to repeat the calculations is presented in Files_GitHub_Mendota2019-2021.pdf.   
These R scripts were written for the manuscript “Stochastic dynamics in years of contrasting phosphorus load” by S.R. Carpenter and W.A. Brock.  
That manuscript explains the methods and cites the original literature that introduced the methods. Correspondence: srcarpen@wisc.edu